pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "docsorter"
version = "0.1.0"
description = "CLI to organize documents by year/month based on dates."
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "Andre Abrantes" }]
requires-python = ">=3.10"
dependencies = [
    "rich>=13.7.0",
    "typer>=0.12.3",
    "PyYAML>=6.0.1",
    "exifread>=3.0.0"
]

[project.scripts]
docsorter = "docsorter.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/docsorter"]
